Updated SecurityConfig to use the new introspector and Renamed AuthZAuthorizationService instance
